[mythtv-cmyth] Recache images on error
For new recording, where the preview hasn't been generated yet,
the image cacher doesn't always recache the file.

So far we're only recaching if we get an empty file from the backend.
Unfortunately we also get sometimes a NULL file (error).
In those cases we didn't recache the file and don't get a preview icon
until the next restart.

This fixes the issue by adding an error count to the jobs
and we recache also on normal errors until a limit is reached.