[cmyth] Align buffer sizes
This adds a buflen parameter to cmyth_livetv_chain_setup to specify
the receive buffer from outside.

In addition this removes setting tcp_buflen to the value returned by
getsockopt(SO_RCVBUF). The reason is that getsockopt will return different
values depending on the platform. Linux systems double the value since
they reserve half of the buffer for metadata. Other systems don't do this
and since that value is reused as input for setsockopt when starting playback,
we create buffers of the doubled size.