[cmyth] Resolve deadlock in cmyth_livetv_chain_switch
Normally "865: ref_release(rec->rec_livetv_file)" never destroys the file.
It decrements only 'ref_count' because the file is hold by the chain itself.
This assumption must be known. If the chain is destroyed from another place,
the file gets destroyed here causing a deadlock (by trying to lock the
control connection for sending DONE to the backend).
It is better to avoid any confusion and destroy the file outside the mutex.

For sample, usage of SET_CHANNEL frees the chain and then on the next file
switch we have a deadlock.