[cmyth] Avoid calls to getnameinfo in cmyth_connect_addr and cmyth_reconnect_addr
Callgrind showed that the connection establishment is slow because of
getnameinfo. The function asks the hostname / user friendly service
names from the system (and in the worst case, from DNS) for printing
the connection string in the log.

To speed up the connection establishment, we use the connection strings
from cmyth_connect, or for cmyth_reconnect_addr: conn->server and
conn->port.