[mythtv-cmyth] EDL: Use position/duration/seeking since MythTV 0.27
Since MythTV 0.27, the backend provides accurate position/duration/seeking with
non-constant frame rates.

Depending of the backend database version, we check if the new type of mark (33)
exists else fallback using old method based on framerate.

Note that to get the benefit of these changes for preexisting recordings,
it may be necessary to run "mythcommflag --rebuild" on affected recordings.