[mythtv-cmyth] When EDL start_mark is zero, assume 0ms offset. Skip DB seek
mythcommflag does not always rebuild the recordedseek table starting from 0,
which can put the start of a break that is at the beginning of a recording
off by a half-second or more.

Also add in a check to see if a DB seek for end_mark offsets failed. If they
failed, fallback to the static framerate guess.