[cmyth] Handle MythTV priorities correctly
MythTV priorities can be significantly larger than the limits of a int8.
This change allows users with recordings that have priorities outside of the
int8 range to load recordings and use the MythTV PVR addon.

Without this, the addon will fail to populate the recordings list if there are
any recordings that had a priority outside the int8 range.