Add Wake On LAN functionality Save/Load server MAC address in custom file in addon_data location
Sends WOL packet to saved MAC address prior to attempting to connect() if
the backend status is unknown or down.
If backend is known to be up, no WOL is sent
Also added debug messages to indicate Response OK and when starting
connection rather than only when it fails
Fixed GetBackendVersion() call, non static CStdString was being cleaned up
invalidating the pointer that was returned from the function.